refactored global search completer into a based and dervied class
