+++ info.yml
@@ -27,5 +27,5 @@ requires_core_version: "v0.14.58"
  
 # the frameworks required to run this app
 frameworks:
-    - {"name": "tk-framework-shotgunutils", "version": "v4.x.x"}
+    - {"name": "tk-framework-shotgunutils", "version": "v5.x.x"}
